# compute: configure DD's idle_merge_effort