@teskje 2022-12-21
I added @philip-stoev because this has the potential to affect compute dataflow performance. The hope is that it doesn't, but I haven't checked yet. I'm planning to do so tomorrow using @vmarcos TPC-H benchmarks.