@teskje 2022-12-22
Unfortunately, both the feature benchmarks and TPC-H suggest that this change degrades performance.

## Feature benchmarks

The feature benchmarks crash at some point (timing out when trying to create a Kafka topic), but before that they output this:

```
NAME                      |    THIS     |    OTHER    |  Regression?  | 'THIS' is:
----------------------------------------------------------------------------------------------------
FastPathFilterNoIndex     |       0.317 |       0.317 |      no       | 0.1 pct   slower
FastPathFilterIndex       |       2.274 |       2.294 |      no       | 0.9 pct   faster
FastPathOrderByLimit      |       0.169 |       0.169 |      no       | 0.2 pct   faster
Insert                    |       3.510 |       2.741 |    !!YES!!    | 28.1 pct   slower
InsertBatch               |       7.296 |       7.314 |      no       | 0.2 pct   faster
InsertMultiRow            |       0.066 |       0.065 |      no       | 1.6 pct   slower
Update                    |       4.909 |       4.777 |      no       | 2.8 pct   slower
UpdateMultiNoIndex        |       5.971 |       5.102 |    !!YES!!    | 17.0 pct   slower
InsertAndSelect           |       4.419 |       3.364 |    !!YES!!    | 31.4 pct   slower
OrderBy                   |       9.234 |       8.788 |      no       | 5.1 pct   slower
CountDistinct             |       0.596 |       0.581 |      no       | 2.6 pct   slower
MinMax                    |       0.701 |       0.697 |      no       | 0.5 pct   slower
GroupBy                   |       3.199 |       3.117 |      no       | 2.6 pct   slower
CrossJoin                 |       1.701 |       1.712 |      no       | 0.7 pct   faster
Retraction                |       2.287 |       2.395 |      no       | 4.5 pct   faster
CreateIndex               |       1.090 |       1.061 |      no       | 2.7 pct   slower
DeltaJoin                 |       0.590 |       0.586 |      no       | 0.7 pct   slower
DifferentialJoin          |       0.528 |       0.476 |    !!YES!!    | 11.0 pct   slower
FullOuterJoin             |       4.734 |       4.908 |      no       | 3.6 pct   faster
FinishOrderByLimit        |       0.485 |       0.471 |      no       | 3.0 pct   slower
```

Some of the timings are significantly slower than on main, which can probably not be attributed to noise in the measurement.

## TCP-H

For TCP-H, I read @vmarcos comment to late and instead come up with the following:

    * Instantiate a TCP-H load generator source with tick interval 1s and scale factor 1.
    * Create PK and FK indexes on all input relations.
    * Create views with default indexes on all TCP-H queries.
    * Check CPU and memory usage.
    
I did all this on an xsmall instance (`default.r1` replica).

This is not a load test, but it includes a lot of arrangements, so I expected it to show the effects of the change in this PR well.

### Results

Setting `idle_merge_effort` does indeed reduce memory usage, but it increases CPU usage massively. This is clearly visible in the Grafana metrics:

* Without `idle_merge_effort`:
    ![Screenshot 2022-12-22 at 13 29 05](https://user-images.githubusercontent.com/4521314/209137581-dbe1d243-c19d-46dd-9394-a0b907bda2fb.png)
* With `idle_merge_effort`:
    ![Screenshot 2022-12-22 at 13 28 06](https://user-images.githubusercontent.com/4521314/209137692-5a507e78-b9f8-484d-b1af-c2304075afaa.png)

Memory usage decreases from 60 to 40%, while CPU usage increases from 6% to 100% (we only allocate half of the logical cores to timely, so the CPU graph is misleading). Looking at the flame graph, the bulk of CPU time is indeed spent in arrangement maintenance:

![Screenshot 2022-12-22 at 13 55 54](https://user-images.githubusercontent.com/4521314/209138910-65b7ec3f-2900-40e6-b19a-ba50dfc58cbb.png)

The jump in CPU usage is way higher than I'd hoped. I didn't test different values for `idle_merge_effort`, it's possible that a smaller one might work out better. In any case, we should probably not merge this PR as-is.