@frankmcsherry 2022-12-22
Would it make sense to offer it as a replica creation argument?