@teskje 2022-12-22
> Would it make sense to offer it as a replica creation argument?

I created an issue to track this here: https://github.com/MaterializeInc/materialize/issues/16794