@teskje 2022-12-22
Unfortunately, an `idle_merge_effort` of 100 also doesn't make the feature benchmarks especially happy:

```
NAME                      |    THIS     |    OTHER    |  Regression?  | 'THIS' is:
----------------------------------------------------------------------------------------------------
FastPathFilterNoIndex     |       0.313 |       0.301 |      no       | 3.9 pct   slower
FastPathFilterIndex       |       2.250 |       2.242 |      no       | 0.4 pct   slower
FastPathOrderByLimit      |       0.170 |       0.169 |      no       | 0.3 pct   slower
Insert                    |       3.742 |       2.808 |    !!YES!!    | 33.3 pct   slower
InsertBatch               |       7.311 |       7.389 |      no       | 1.1 pct   faster
InsertMultiRow            |       0.065 |       0.066 |      no       | 2.5 pct   faster
Update                    |       4.928 |       4.700 |      no       | 4.8 pct   slower
UpdateMultiNoIndex        |       5.997 |       5.091 |    !!YES!!    | 17.8 pct   slower
InsertAndSelect           |       4.335 |       3.391 |    !!YES!!    | 27.8 pct   slower
OrderBy                   |       8.407 |       8.776 |      no       | 4.2 pct   faster
CountDistinct             |       0.649 |       0.609 |      no       | 6.5 pct   slower
MinMax                    |       0.705 |       0.703 |      no       | 0.4 pct   slower
GroupBy                   |       3.191 |       3.173 |      no       | 0.6 pct   slower
CrossJoin                 |       1.857 |       1.825 |      no       | 1.7 pct   slower
Retraction                |       2.288 |       2.386 |      no       | 4.1 pct   faster
CreateIndex               |       1.088 |       1.072 |      no       | 1.5 pct   slower
DeltaJoin                 |       0.604 |       0.598 |      no       | 1.0 pct   slower
DifferentialJoin          |       0.520 |       0.491 |      no       | 6.0 pct   slower
FullOuterJoin             |       5.207 |       4.901 |      no       | 6.3 pct   slower
FinishOrderByLimit        |       0.468 |       0.482 |      no       | 2.8 pct   faster
```