@frankmcsherry 2022-12-22
Hrm. I wonder if the benchmarks are just a bit weird / pessimal. The "insert" benchmarks seem likely to trickle in changes, which will cause a relatively larger amount of CPU to be used merging. Mysterious! I'm surprised that inserting takes more time, but I suspect we are measuring more than just insertions (also reading it back, maybe?).