@teskje 2023-01-02
Thanks Philip!

As suggested, I did some more TPC-H benchmarking. This time I created a one-off TPC-H load generator and measured the time it took to run the queries Q01 to Q22 in sequence. I did this with no indexes and with the full set of pk/fk indexes.

Here are the results (times are minimum values of two runs each):

* Without indexes:

  |    | idle_merge_effort = None | idle_merge_effort = 1000 |
  |---|---|---|
  |  total query time | 266 s | 255 s | 
  |  max memory usage | 2.16 GB | 2.56 GB |
  |  max CPU usage | 80% | 83% |

* With indexes:

  |    | idle_merge_effort = None | idle_merge_effort = 1000 |
  |---|---|---|
  |  total query time | 42 s | 37 s | 
  |  max memory usage | 7.17 GB | 7.78 GB |
  |  max CPU usage | 80% | 100% |

Queries under idle_merge_effort are faster than queries without idle_merge_effort. The difference is not huge, but it is consistent from what I can tell, also between individual TPC-H queries.

As expected, setting idle_merge_effort leads to higher CPU usage.

What's unexpected is that setting idle_merge_effort also leads to higher memory usage. We would expect to see the opposite. But then again the differences are not huge and I simply eyeballed these metrics from Grafana, so it is entirely possible that this is just noise. My previous TPC-H benchmark showed that there can be a significant reduction in memory usage from enabling idle_merge_effort.

Based on these results, and my previous smooth-rehydration experiments, it does seem like setting idle_merge_effort does have a benefit. So I think it's fine to merge this PR, but I'm also happy to do further tests if anyone would like to suggest some!